requirements.txt
asgiref==3.2.10
attrs==20.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
Django==3.1
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-translate==3.0.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
libcst==0.3.10
murmurhash==1.0.2
mypy-extensions==0.4.3
mysqlclient==2.0.1
numpy==1.19.1
plac==1.1.3
preshed==3.0.2
proto-plus==1.9.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.1
tqdm==4.48.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
wasabi==0.8.0
